Green LED addition!!!
Edited the MAX30101.h and MAX30101.c and main.c file. Firstly, updated the main.c to add the twi_master_init function to begin the TWI communications. This fixed the communication errors to the sensor. Then we added functionality for Multi-LED mode in the sensor. This required adding additional functionality in the MAX30101.h file, updating the .c file to include the 3rd LED and then pass that Green LED to the heart rate algo.